reorganized models into sub directories
