+++ src/config/database.ts
@@ -1,10 +1,14 @@
 import { createConnection, ConnectionOptions, Connection } from "typeorm";
 import Business from "../models/business";
 import Department from "../models/department";
+import Manual from "../models/manual/manual";
+import Policy from "../models/manual/policy";
+import Section from "../models/manual/section";
+import ManualAssignment from "../models/manual/manual_assignment";
 import Permission from "../models/permission";
 import Role from "../models/role";
-import User from "../models/user";
-import UserRole from "../models/user_role";
+import User from "../models/user/user";
+import UserRole from "../models/user/user_role";
 import DBLogger from "../util/logs/db_logger";
 
 const connection: ConnectionOptions = {
@@ -15,7 +19,18 @@ const connection: ConnectionOptions = {
     // enforce strict typing by only applying
     // a small subset of the potential database types
     type: (process.env.DB_TYPE as "mysql" | "postgres") ?? "",
-    entities: [Business, User, Department, Permission, Role, UserRole],
+    entities: [
+        Business,
+        User,
+        Department,
+        Permission,
+        Role,
+        UserRole,
+        Manual,
+        ManualAssignment,
+        Section,
+        Policy,
+    ],
     logging: true,
     logger: new DBLogger(),
 };
